requirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
dm-tree==0.1.9
filelock==3.17.0
Flask==3.1.0
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.2.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
humanfriendly==10.0
idna==3.10
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
keras==3.9.1
kiwisolver==1.4.8
langdetect==1.0.9
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
oauthlib==3.2.2
onnx==1.16.2
onnx-simplifier==0.4.36
onnx-tf==1.10.0
onnx2tf==1.22.3
onnx_graphsurgeon==0.5.6
onnxruntime==1.21.0
onnxslim==0.1.48
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
Pygments==2.19.1
pyparsing==3.2.1
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-engineio==4.12.0
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scipy==1.15.2
seaborn==0.13.2
simple-websocket==1.1.0
six==1.17.0
sng4onnx==1.0.4
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow-addons==0.22.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.25.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf2onnx==1.16.1
tflite-support==0.1.0a1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.82
ultralytics-thop==2.0.14
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
